Delete account with password confirmation

+++ jcasts/users/views.py
@@ -16,7 +16,7 @@ from django.views.decorators.http import require_http_methods
 
 from jcasts.episodes.models import AudioLog, Bookmark
 from jcasts.podcasts.models import Podcast, Subscription
-from jcasts.users.forms import UserPreferencesForm
+from jcasts.users.forms import UserDeleteForm, UserPreferencesForm
 
 
 @require_http_methods(["GET", "POST"])
@@ -130,12 +130,17 @@ def user_stats(request: HttpRequest) -> HttpResponse:
 @require_http_methods(["GET", "POST"])
 @login_required
 def delete_account(request: HttpRequest) -> HttpResponse:
-    if request.method == "POST" and "confirm-delete" in request.POST:
+    form = UserDeleteForm(request.user, request.POST or None)
+    if (
+        request.method == "POST"
+        and form.is_valid()
+        and "confirm-delete" in request.POST
+    ):
         request.user.delete()
         logout(request)
         messages.info(request, "Your account has been deleted")
         return HttpResponseRedirect(settings.HOME_URL)
-    return TemplateResponse(request, "account/delete_account.html")
+    return TemplateResponse(request, "account/delete_account.html", {"form": form})
 
 
 def with_export_response_attachment(
